Reconcile the 22 gplates-only commits into pygplates
Applies the same tree-level resolution as the original merge of
'gplates' (tip b4fa68e49) into this pygplates-derived staging branch,
but recorded as a regular single-parent commit rather than a git merge
commit. gplates and pygplates remain separate long-lived develop
branches with their own release cadence; this keeps the staging
branch's history mergeable as a plain PR into pygplates (not pygplates
gplates), and pygplates can still be merged into gplates later at any
time without needing to cherry-pick.

Brings the pygplates lineage up to date with (post-2.5 fixes):
- GPlates app:      cmake -DGPLATES_BUILD_GPLATES=TRUE (the default)
- pyGPlates module: cmake -DGPLATES_BUILD_GPLATES=FALSE

Conflict resolutions (per pygplates/MERGE-PLAN.md):
- Version.cmake: GPlates 2.6.0-4, pyGPlates 1.1.0.dev5, unified header
  comment, and GPLATES_BUILD_GPLATES option default flipped to true
  (the auto-merge would have silently kept pygplates' false default).
- src/CMakeLists.txt: Boost min version 1.69, no boost::system component.
- ScribeExportPyGPlates.cc: keep model/property-values scribe exports.
- ReconstructionGeometryRenderer.cc: keep gplates' file-local
  get_subduction_polarity() helper (drop pygplates' orphaned member
  function) and port its 3 call sites to the revisioned model API
  (type()->get_type(), value()->get_value()).
- InstallSharedLibraryDependencies.cmake: keep pygplates' Python-stdlib
  install-prefix injection plus gplates' generalized all-frameworks
  macOS codesign loop.
- python.qrc: union (hellinger.py + the 7 python/api/*.py resources).
- Deleted src/api/Python.cc and src/api/PyFunctions.cc: the embedded
  Python console now uses the full pygplates API via PyGPlatesModule.cc
  (old thin wrappers remain as OldFeature/OldFeatureCollection;
  reconstruct/reverse_reconstruct have the new-API signatures).
- BUILD.*: unified wording (one source tree builds both products).
- Untangled spurious cstdint.hpp rename/rename (kept VelocityUnits.h
  and PyInformationModel.cc; cstdint.hpp stays deleted).

Provenance: original two-parent merge commit was 1b178f7e3 (parents
69a58bbc0, b4fa68e49); superseded by this non-merge commit.
