+++ BUILD.OSX
@@ -15,8 +15,9 @@ Introduction
 
 This is a guide to building GPlates or pyGPlates from source code in a macOS development environment.
 
-Note: This source code originated from a "pygplates" branch of the GPlates GitHub repository
-      (https://github.com/GPlates/GPlates) and should only be used to compile pyGPlates (not GPlates).
+Note: This source code can be used to compile either GPlates or pyGPlates (but not both in the same build).
+      By default GPlates is built. To build pyGPlates instead, set the CMake variable GPLATES_BUILD_GPLATES
+      to FALSE (eg, cmake -DGPLATES_BUILD_GPLATES=FALSE ...). Setting it to TRUE (the default) builds GPlates.
 
 
 Development environment
