+++ src/qt-resources/python.qrc
@@ -3,7 +3,9 @@
     <file>python/scripts/colouring/ArbitraryColours.py</file>
     <file>python/scripts/colouring/ColorByProperty.py</file>
     <file>python/scripts/colouring/draw_style_demo.py</file>
-    
+
+    <file>python/scripts/hellinger/hellinger.py</file>
+
     <file>python/api/Crossovers.py</file>
     <file>python/api/Feature.py</file>
     <file>python/api/GeometriesOnSphere.py</file>
